+++ changelog.md
@@ -1,10 +1,12 @@
-# Changes in 1.16
+# Changes in 1.17
 
 ## Breaking changes
 
 ## New features
 
 ## Bug fixes
+
+* Fix reporting error based on http responses from the Aura-API with an invalid JSON body. Earlier the client would report JSONDecodeError instead of showing the actual issue.
 
 ## Improvements